extract header into own component
